requirements.txt
langchain_openai
openai
python-dotenv
streamlit